Add top-level subType property to segment (~kind from PR#10)
This property is equivalent to the "kind" property in PR#10. I renamed
to subType to try to reduce the number of terms floating around.

When subType=road, then the property "road" may be defind within the
feature's top-level properties. The schema is designed so that you
can omit "road" or any of its sub-members and still havea a reasonable
road feature based on the defaults.

See also: https://github.com/OvertureMaps/schema-wg/pull/10