Allow speedLimit rules to be constrained to specific lanes
I did not add a lanes constraint to the maximum size rules because,
for now, it's not clear how badly it is needed. The relevant OSM tags
do not explicitly mention lanes variants, with the exception of one
offhand mention for maxweight.

- https://wiki.openstreetmap.org/wiki/Key:maxwidth
- https://wiki.openstreetmap.org/wiki/Key:maxweight
- https://wiki.openstreetmap.org/wiki/Key:maxheight

Taginfo indicates there are small numbers of OSM tags out there
with lane-specific info (low hundreds):

_ https://taginfo.openstreetmap.org/keys/maxwidth:lanes
_ https://taginfo.openstreetmap.org/keys/maxweight:lanes
- https://taginfo.openstreetmap.org/keys/maxheight:lanes

Versus 16K ways with maxspeed:lanes.