First item in standard names must be "local" locale
To make things somewhat easier for users, the first element
in the standard name must have the "local" locale.
This makes things easier because the local locale is the easiest
"default" to use. And because:
* The standard name is required for any "names" field
* The standard name is required to have at least one item
* The first item is required to be the "local" locale
any feature with a "names" property can be easily used
by taking the first element of the standard name

Note that this uses the `$ref: "..."` syntax because the
{ "$ref": ... } syntax used elsewhere was failing as invalid yaml
so this seems to work.