Support expected error assertions within the counterexamples
This change enables us to assert that a particular counterexample
results in a particular validation error, thus preventing the build-up
of cruft from pointless counterexamples which aren't testing anything
of value.

As part of this change, I had to delete the `time-no-timezone.yaml`
example because it was an invalid example that had been silently
missed previously but started failing when I added the `-assertformat`
assertion requirement to the `jv` command line.