Improve and standardize additive sub-schema descriptions and titles
The reason for this change is that in the Docusaurus documentation,
the browsable version of the schema includes both the title (at the
top of the browsable control) and the description (at the bottom of
the control). The previous titles and descriptions were confusing
and failed to give much useful information.