Convert property names and enumeration member names to snake_case
This commit switches properties and enumerators from `camelCase` naming
to `snake_case` naming.

The purpose is to improve compatibility with SQL based tools. Because
SQL uses case-insensitive identifiers for historical reasons, SQL-based
tools cause problems and information loss with camel-casing because they
tend to treat all identifiers as if they were alllowercase or ALLCAPS.

In the Overture Schema Task Force meeting of 2024-02-07, the task force
decided to adopt the snake_case proposal:

https://github.com/OvertureMaps/schema-wg/issues/272

The proposal originated in Jake's Schema Friction discussion document
on the Overture Confluence:

https://wiki.overturemaps.org/x/SQGAAQ

While we believe that snake_case is slightly unconventional, and dare I
(Vic) say also more unsightly in the JSON context, we feel that for SQL
compatibility reasons, the pros of changing outweigh the cons.