Use the feature's type as its title for each feature type sub-schema
Some code generation tools (POJO, POPO, etc.) use the JSON Schema
`title` as a hint for naming the object that represents the schema.

Thus this change increases the probability that, for example, if you
use a code generation tool to generate a Plain Old Python Type for
the `segment` feature type, it will be named `segment`.