Correct scoping implementation at schema level
This is an omnibus commit addressing mainly scoping bugs that affect the
transportation schema. These are bugs where the schema was wrong and did
not accurately reflect either the design intention, or the way the data
were being generated/combobulated, or both.

As part of this commit, an effort has been made to bring the terminology
used in the schema into line with the terminology used in the design
documents and the transportation combobulator.

The full listing of issues that are fully fixed by this commit is below:

1. [Bug] Lanes model supplies default LR range at point of use:
   https://github.com/OvertureMaps/tf-transportation/issues/128. This
   bug is fully fixed by removing the default value.

2. [Bug]: Lane block rules do not support temporal scoping (schema
   only): https://github.com/OvertureMaps/tf-transportation/issues/127.
   This bug is fully fixed by adding support for temporal scoping at
   the lane block rule level.

3. [Bug]: Remove mixed schema types to fix Parquet and codegen
   compatibility 4 of 4: lanes:
   https://github.com/OvertureMaps/tf-transportation/issues/124. This
   bug is fully fixed by deleting the flat lane block option and
   always requiring a list of lane block rules.

4. [Bug]: Speed limits schema scoping support is broken:
   https://github.com/OvertureMaps/tf-transportation/issues/113. This
   bug is fully fixed by pushing travel mode scopes down under `when`
   and adding support for heading scope for speed limits, and also
   adding all the other known relevant scopes at the same time,
   including vehicle property scoping most importantly.

5. [Bug]: Missing units and vehicle properties in vehicle scoping -
   length, width, height, and weight:
   https://github.com/OvertureMaps/tf-transportation/issues/87. This
   bug is fully fixed by correctly modeling vehicle property scopes.

In addition, this commit starts paving the way toward implementing
`except` (https://github.com/OvertureMaps/tf-transportation/issues/111)
for scoping. It does so by getting rid of `mode_not`, which is currently
unused and will be rendered obsolete by `except`.