Push segment level down into road property, add linear referencing
This commit fixes the schema part of the following issue:

    "[Bug]: Level missing 84% of eligible segments (and is in the wrong
            place)"
        https://github.com/OvertureMaps/tf-transportation/issues/126

The combobulation aspect of the fix can be found at:

    "Level down and class up"
        https://github.com/OvertureMaps/tf-transportation/pull/162

The fix is, firstly, to convert the segment `level` property from a
scalar integer valued property to a rule-based property where the
rules support only geometric scope (linear referencing).

Because the rules/LR fix makes the `level` property incompatible with
the basic scalar-valued property used in buildings, the second part
of the fix is to push `level` down under the `road` property.

Note: When we start to fill out `rail` properties, we will likely
      create a `rail` property and level for railroads will be
      found nested underneath it, analogous to roads.