Initial addresses schema
We recently decided that the initial addresses schema would be based on the
output of the Open Addresses (OA) "conform" schema. The definition of the fields
can be found at
https://github.com/openaddresses/openaddresses/blob/master/schema/layers/address_conform.json

This initial schema is a strawman intended for feedback and discussion.

Our initial schema looks at all properties included in the US Northeast distribution
which includes the following fields:
* id
* country (the first element in the filepath e.g. us/ma/city_of_boston-addresses-city.geojson is "us")
* subcountry (the second element in the filepath e.g. us/ma/city_of_boston-addresses-city.geojson is "ma") - This field should be renamed or be removed
* region
* district
* city
* postcode
* street
* number
* unit