Align boundary division IDs with schema relationships guidance
In a recent change, https://github.com/OvertureMaps/schema/pull/200,
the schema for the divisions theme's boundary feature fell out of
alignment with the *Relationships in the Schema* guidance.

The guidance is that a property or nested sub-property that contains
a foreign key reference to another feature by Overture ID must end
in the suffix `_id`.

Before PR #200, things were fine because the Overture ID was contained
in a nested sub-property, `divisions[*].division_id`. But since PR #200
removed the nested structures from the `divisions` property, it should
have renamed it `division_ids`. This change fixes that.

References:

- https://github.com/OvertureMaps/schema-wg/blob/admin/decisions/adr-relationship-modeling.md
- https://wiki.overturemaps.org/x/oAEjAQ