Allow example files to be JSON
For some reason, at some point, we decided to force all example files to
be YAML. But since Dana wants to have a JSON example on the main schema
documentation page, and since we want to write it "natively" in JSON so
we can control the formatting, this choice seems to be obsolete.

This change converts the "examples must be YAML" rule to an "examples
must be YAML or JSON rule". It also fixes a bug where the rule wasn't
being enforced.