Validate place categories are snake_case identifiers
This change ensures that place categories meet our basic criteria for
being Python-style snake_case identifiers.

As a result of this change, 861,742 place features from the
v2024_05_16_beta_0 release would fail schema validation. That is
because there are 6 lines in `task_force_docs/places/overture_categories.csv`
that match the regular expression /[^a-z0-9_;\[\],\s]/. These are:

    69: hong_kong_style_cafe;[eat_and_drink,restaurant,asian_restaurant,hong_kong_style_café]
   483: children's_museum;[attractions_and_activities,museum,children's_museum]
   873: children's_clothing_store;[retail,shopping,fashion,clothing_store,children's_clothing_store]
   881: men's_clothing_store;[retail,shopping,fashion,clothing_store,men's_clothing_store]
   885: women's_clothing_store;[retail,shopping,fashion,clothing_store,women's_clothing_store]
  1224: women's_health_clinic;[health_and_medical,women's_health_clinic]

I will file an issue to fix this.