Use element_at() instead of subscript for tags map key lookup
When retrieving a key from a map (like tags in OSM), our queries typically use
subscript syntax like tags['highway']. The default behavior of subscript syntax
is to fail and error if 'highway' is not in the tags map. However, there is a
session property in trino and presto that overrides this behavior allowing it
to return null.  It seems that Athena and other systems are deployed with this
behavior which makes it somewhat non-standard usage. Instead, we should now use
element_at(tags, 'highway') which explicitly returns null when 'highway' is not
in the tags map. The behavior should be unchanged when Trino/Presto is
configured with the legacy behavior, but will have identical results when those
systems are deployed with the more modern defaults.