Add class to divisions schema (#301)
This change adds a `class` property to the division feature.

The intention is to enable further differentiation of the "locality" `subtype`
in places where we don't currently have population data. (There may also be
an intention to apply the `class` labels differently depending on local factors
e.g. the standard for a megacity might possibly be different in China versus
Mongolia.)

Currently, only the only supported `class` labels are for the "locality"
division `subtype`, but it is of course possible that `class` will be
extended to cover other `subtype` values in the future.