Delete misleading `FeatureUpdateTime` type
This type was unnecessary and misleading.

The misleading part comes from the fact that, historically, the Overture
feature in JSON Schema land had an `update_time` property. We eventually
removed that property, making a reusable type called `FeatureUpdateTime`
wholly redundant, but it seems the type was reused by the `update_time`
property in the sources construct.

This lead to misleading-ness not only at the type name/intent level, but
also at the documentation level since the documentation inherited by the
update time field in sources was:

> Timestamp when the feature was last updated

Which assuredly made no sense for sources. As a result of this commit,
that documentation is now updated to:

> Last update time of the source data record

Which makes more sense (and hopefully is also correct).