Fix bug: Incorrect defaults (and eliminate all Pydantic warnings) 🐞
The biggest subtle bug was that `level_rules` in the transportation
segment were ending up with a default value of `0`, so that you could
just omit `value` altogether. This was not intended as it's bizarre,
better to not create the rules at all. Meanwhile, `level` in all the
other features did NOT have a default of zero, which was again not
intended.

As a result of the `Stacked` model overriding the default value that
is set on `Level` in its `level` field, we were also getting a Pydantic
warning that was polluting the test run output. By fixing the underlying
issue, this commit makes that warning go away.