Migrate cartography into is own module
This commit also fixes one definite bug, and one weird thing that was
arguably a bug.

Definite bug: `Prominence` was modeled as `lt=100`, which would
indicate a peculiar-sized 99-element range from 1 to 99 inclusive, but
the actual data in the latest Overture release, `v2025_10_22_0`, ranges
from 1 to 100 inclusive.

Weird thing: Sort key had a default of zero, and zero is also supposed
to be the highest rendering priority. This seems weird. It would mean
that you can't use the sort key to give any feature higher rendering
priority than a feature that has no rendering priority set. This won't
work as expected, so I removed the default.

To fix another day: prominence and sort key are reversed. A high number
means high prominence, but a high number for sort key^[ means a low
rendering priority.