Fix majority of `pydocstyle` errors
As of this commit, running `make docformat` detects there are 22
remaining errors, less than one terminal full. They are all of
the following two types:

1. D100: Missing docstring in public module
2. D101: Missing docstring in public class

Moreover all the remaining errors pertain to modules and/or classes that
are slated for further refactoring. Therefore, I'm going to leave
finishing them for another day.

As part of this commit, I edited the `make docformat` rule in the
`Makefile` to suppress four new rules. The rules suppressed are
listed below along with the description of the rule and the reason for
suppressing it.

1. D102: Missing docstring in public method. This is suppressed because
   `pydocstyle` is dumb and doesn't traverse inheritance hierarchies
   or even look at the `@override` decorator, so it needlessly flags
   overridden methods for which documentation is superfluous and a
   waste of time.
2. D200: One-line docstring should fit on one line with quotes. This is
   suppressed because `pydocstyle` is dumb and isn't aware of reasonable
   line length limits or wrapping.
3. D205: 1 blank line required between summary line and description.
   This is suppressed for the same reason as D200.
4. D400: First line should end with a period. This is suppressed for the
   same erason as D200 and D205.