Remove tool: `docformatter`
The issues are that `docformatter` doesn't understand NumPy-style
documentation, which is how most of the API docs are written today, and
that it edits code in the opposite way to what `pydocstyle` expects.

Both `docformatter` and `pydocstyle` are somewhat underwhelming tools
with very few configuration options and a very heavy-handed approach
to what they do, but as `pydocstyle` understands more documentation
"dialects", including NumPy, it seems reasonable to keep it and drop
the other, at least for now.