Fix bug: `OvertureFeature` overrode and broke `Feature` GeoJSON parse 🦟
This was a subtle bug. Only because we weren't previously exercising this
functionality at the `OvertureFeature` level via unit test.

Basically the Pydantic decorators like `@model_validator` break in an
inheritance context if the derived class overrides a validator method
that was used in a base class. This only makes sense and isn't really
Pydantic's fault: it'll be trying to call `instance.validator_method()`
and Python is going to resolve which code to invoke. If you overrode
the base class' `validator_method()` then yes, it stands to reason
that only the derived version will be called.

As you can see from the diff, the wrap validator in `Feature` that had
been fixing up the input data for GeoJSON conversion was called
`validate_model`, and I had created a method of the same name in
`OvertureFeature` to handle the soon-to-be-deprecated `ext_*` field
exception. So, of course, `Feature`'s version got overridden, and with
it the whole GeoJSON-to-flat-model transform that enabled the parsing
to continue.

I fixed the issue by giving the method names in both classes much more
specific names and tried to make them a bit obscure with dunder-naming,
which won't obfuscate them but will at least hopefully reduce the
chances of name collisions.

Also added a test, because the lack of it is the cause of this whole
fiasco in the first place!